improve -Organize the binary-tree sources.
