+++ data-structure/tree/binary-tree/include/binary_tree.h
@@ -1,26 +1,24 @@
 #ifndef BINARY_TREE
 #define BINARY_TREE
 
-typedef struct node {
-	int data;
-	struct node *left;
-	struct node *right;
-} Node;
+typedef struct tree_node {
+    /* the structure for a binary tree */
+
+    int data;
+    struct tree_node *left;  // pointer to the left child
+    struct tree_node *right; // pointer to the right child
+} TreeNode;
 
 
-Node *initializeNode(int value);
+TreeNode *initializeTreeNode(int value);
 
-void deleteTree(Node **node);
+void deleteTree(TreeNode **node);
 
-int heightNode(Node *node);
-int maxNode(Node *node);
-unsigned int leafQuantity(Node *node);
-unsigned int nodesQuantity(Node *node);
+int heightNode(TreeNode *node);
+int maxNode(TreeNode *node);
+unsigned int leafQuantity(TreeNode *node);
+unsigned int nodesQuantity(TreeNode *node);
 
-
-void preorder(Node *node);
-void inorder(Node *node);
-void postorder(Node *node);
 
 
 void run_binary_tree_example();
