+++ data-structure/tree/binary-tree/include/stack.h
@@ -1,0 +1,21 @@
+#ifndef STACK_H
+#define STACK_H
+
+#include "traverse.h"   // use for get the TreeNode structure
+
+
+typedef struct linked_list_node {
+    /* the structure for a node in the linked list stack */
+
+    TreeNode *tree_node;    // pointer to a tree node (data stored in the stack)
+    struct linked_list_node *next;  // pointer to the next node in the stack
+} StackNode;
+
+
+StackNode *initializeStackNode(TreeNode *tree_node);
+bool isEmpty(StackNode *top);
+void push(StackNode **top, TreeNode *node);
+TreeNode *pop(StackNode **top);
+
+
+#endif  // STACK_H
