+++ data-structure/tree/binary-tree/include/traverse.h
@@ -1,0 +1,12 @@
+#ifndef TRAVERSE_BINARY_TREE
+#define TRAVERSE_BINARY_TREE
+
+#include "binary_tree.h"    // use for get the Node structure
+
+
+void inorder_recursive(TreeNode *node);
+void preorder_recursive(TreeNode *node);
+void postorder_recursive(TreeNode *node);
+
+
+#endif  // TRAVERSE_BINARY_TREE
